SpriteGraphs can have Attachments added to them.
SpriteGraph attachments are just nodes which contain render information,
used for rendering a particular node. The information consists of a
SpriteSheetDefinitionName and DefaultSpriteName.
Sprite Graph node instances will, be rendered, first attempt to find a
sprite within the spritesheetdefinition which matches the current
SpriteGraph animation. Failing that, it will attempt to load the sprite
with the 'default spritename'.